Add pthread_barrier event
This is just like the barrier event except it actually uses
pthread_barrier_wait instead of pthread_cond_wait and
pthread_cond_broadcast. This alternative method may cause issues with
killing rt-app, but is still desirable because it exercises different
syscall usage patterns, which are used in real workloads. I have
tried to explain the details of this in tutorial.txt

Implementing this requires adding a `finalize` step to resource
setup, because we need to know the number of threads that refer to
the resource before we can call pthread_barrier_init.